+++ lib/i5_supreme.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require_relative "i5_supreme/version"
+require_relative "i5_supreme/paths"
+require_relative "i5_supreme/manifest_manager"
+require_relative "i5_supreme/rule_auditor"
+require_relative "i5_supreme/release_pipeline"
+require_relative "i5_supreme/cli"